Remove obsolete iOS sync setup UI
